more progress on posix API layer
see #2380
