+++ CMakeLists.txt
@@ -474,6 +474,10 @@ set(ZIG_STD_FILES
     "c/linux.zig"
     "c/netbsd.zig"
     "c/windows.zig"
+    "c/posix.zig"
+    "c/posix/darwin.zig"
+    "c/posix/freebsd.zig"
+    "c/posix/windows.zig"
     "coff.zig"
     "crypto.zig"
     "crypto/blake2.zig"
@@ -612,15 +616,20 @@ set(ZIG_STD_FILES
     "os/linux.zig"
     "os/linux/arm64.zig"
     "os/linux/errno.zig"
+    "os/linux/posix.zig"
+    "os/linux/posix/arm64.zig"
+    "os/linux/posix/x86_64.zig"
     "os/linux/tls.zig"
     "os/linux/vdso.zig"
     "os/linux/x86_64.zig"
     "os/netbsd.zig"
     "os/netbsd/errno.zig"
     "os/path.zig"
+    "os/posix.zig"
     "os/time.zig"
     "os/uefi.zig"
     "os/wasi.zig"
+    "os/wasi/posix.zig"
     "os/windows.zig"
     "os/windows/advapi32.zig"
     "os/windows/errno.zig"
