--- std/os/windows/errno.zig
@@ -1,1 +1,0 @@
-// TODO get these values from msvcrt
